+++ test/db/postgresql/validation_test.dart
@@ -5,7 +5,7 @@ import '../../helpers.dart';
 void main() {
   ManagedContext ctx;
   setUpAll(() async {
-    ctx = await contextWithModels([T]);
+    ctx = await contextWithModels([T, U, V, W]);
   });
 
   tearDownAll(() async {
@@ -117,6 +117,47 @@ void main() {
     expect(o.aOrb, "c");
     expect(o.equalTo1OnInsert, 10);
   });
+
+  test("willUpdate runs prior to update", () async {
+    var q = new Query<U>();
+    var o = await q.insert();
+    q = new Query<U>()..where.id = o.id;
+    o = (await q.update()).first;
+    expect(o.q, "willUpdate");
+  });
+
+  test("willInsert runs prior to insert", () async {
+    var q = new Query<U>();
+    var o = await q.insert();
+    expect(o.id, isNotNull);
+    expect(o.q, "willInsert");
+
+  });
+
+  test("ManagedObject onUpdate is subject to all validations", () async {
+    var q = new Query<V>();
+    var o = await q.insert();
+
+    q = new Query<V>()..where.id = o.id;
+    try {
+      await q.update();
+      expect(true, false);
+    } on QueryException catch (e) {
+      expect(e.toString(), contains("'f'"));
+      expect(e.toString(), contains("Must be one of"));
+    }
+  });
+
+  test("ManagedObject onInsert is subject to all validations", () async {
+    var q = new Query<W>();
+    try {
+      await q.insert();
+      expect(true, false);
+    } on QueryException catch (e) {
+      expect(e.toString(), contains("'f'"));
+      expect(e.toString(), contains("Must be one of"));
+    }
+  });
 }
 
 class T extends ManagedObject<_T> implements _T {}
@@ -135,4 +176,52 @@ class _T {
   @Validate.compare(equalTo: 2, onUpdate: true, onInsert: false)
   @ManagedColumnAttributes(nullable: true)
   int equalTo2OnUpdate;
+}
+
+class U extends ManagedObject<_U> implements _U {
+  @override
+  void willUpdate() {
+    q = "willUpdate";
+  }
+
+  @override
+  void willInsert() {
+    q = "willInsert";
+  }
+}
+class _U {
+  @managedPrimaryKey
+  int id;
+
+  String q;
+}
+
+class V extends ManagedObject<_V> implements _V {
+  @override
+  void willUpdate() {
+    f = "c";
+  }
+}
+class _V {
+  @managedPrimaryKey
+  int id;
+
+  @Validate.oneOf(const ["a", "b"])
+  @ManagedColumnAttributes(nullable: true)
+  String f;
+}
+
+class W extends ManagedObject<_W> implements _W {
+  @override
+  void willInsert() {
+    f = "c";
+  }
+}
+class _W {
+  @managedPrimaryKey
+  int id;
+
+  @Validate.oneOf(const ["a", "b"])
+  @ManagedColumnAttributes(nullable: true)
+  String f;
 }